### 2023-07-19 #5667 @editorialbot
[PRE REVIEW]: automata: A Python package for simulating and manipulating automata
**Submitting author:** @eliotwrobson (<a href="http://orcid.org/0000-0002-1476-6715">Eliot Wong Robson</a>)
**Repository:** https://github.com/caleb531/automata
**Branch with paper.md** (empty if default branch): develop
**Version:** v8.0.0
