### 2023-07-19 #5667 @danielskatz
@eliotwrobson - thanks for your submission. I'll now work on finding an editor. I was going to do this myself, but then saw the institutional conflict I have with you.

Please fix the invalid DOIs that editorialbot found. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.

In addition, please provide suggestions for potential reviewers by mentioning them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
