+++ joss/paper.bib
@@ -1,20 +1,15 @@
-@inproceedings{Sutner02a,
-  author       = {Klaus Sutner},
-  editor       = {Jean{-}Marc Champarnaud and
-                  Denis Maurel},
-  title        = {automata, a Hybrid System for Computational Automata Theory},
-  booktitle    = {Implementation and Application of Automata, 7th International Conference,
-                  {CIAA} 2002, Tours, France, July 3-5, 2002, Revised Papers},
-  series       = {Lecture Notes in Computer Science},
-  volume       = {2608},
-  pages        = {221--227},
-  publisher    = {Springer},
-  year         = {2002},
-  url          = {https://doi.org/10.1007/3-540-44977-9\_21},
-  doi          = {10.1007/3-540-44977-9\_21},
-  timestamp    = {Thu, 02 Jun 2022 15:21:24 +0200},
-  biburl       = {https://dblp.org/rec/conf/wia/Sutner02a.bib},
-  bibsource    = {dblp computer science bibliography, https://dblp.org}
+@InProceedings{Sutner03,
+author="Sutner, Klaus",
+editor="Champarnaud, Jean-Marc
+and Maurel, Denis",
+title="automata, a Hybrid System for Computational Automata Theory",
+booktitle="Implementation and Application of Automata",
+year="2003",
+publisher="Springer Berlin Heidelberg",
+address="Berlin, Heidelberg",
+pages="221--227",
+abstract="We present a system that performs computations on finite state machines, syntactic semigroups, and one-dimensional cellular automata.",
+isbn="978-3-540-44977-5"
 }
 
 @Misc{brics,
