### 2023-07-19 7f0b31a
Update paper reference (#165)
