+++ joss/paper.bib
@@ -1,15 +1,16 @@
 @InProceedings{Sutner03,
-author="Sutner, Klaus",
-editor="Champarnaud, Jean-Marc
-and Maurel, Denis",
-title="automata, a Hybrid System for Computational Automata Theory",
-booktitle="Implementation and Application of Automata",
-year="2003",
-publisher="Springer Berlin Heidelberg",
-address="Berlin, Heidelberg",
-pages="221--227",
-abstract="We present a system that performs computations on finite state machines, syntactic semigroups, and one-dimensional cellular automata.",
-isbn="978-3-540-44977-5"
+  author={Sutner, Klaus},
+  editor={Champarnaud, Jean-Marc and Maurel, Denis},
+  title={automata, a Hybrid System for Computational Automata Theory},
+  booktitle={Implementation and Application of Automata},
+  year={2003},
+  publisher={Springer Berlin Heidelberg},
+  address={Berlin, Heidelberg},
+  pages={221--227},
+  abstract={We present a system that performs computations on finite state machines, syntactic semigroups, and one-dimensional cellular automata.},
+  isbn={978-3-540-44977-5},
+  url={https://doi.org/10.1007/3-540-44977-9_21},
+  doi={10.1007/3-540-44977-9_21},
 }
 
 @Misc{brics,
