### 2023-07-19 4696100
Add internal minify call to to_partial function (#164)
