### 2023-07-22 e9556d2
Update Publish CI job to use PyPI Trusted Publishers feature
This allows for GitHub Actions to authenticate with PyPI directly,
rather than requiring us to manually manage API tokens.
