+++ .github/workflows/publish.yml
@@ -12,6 +12,10 @@ on:
 jobs:
   publish:
     runs-on: ubuntu-latest
+    environment: release
+    # Required for PyPI Trusted Publishers feature
+    permissions:
+      id-token: write
 
     steps:
     - uses: actions/checkout@v3
@@ -40,11 +44,9 @@ jobs:
     - name: Publish distribution to Test PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
-        password: ${{ secrets.TEST_PYPI_API_TOKEN }}
         repository_url: https://test.pypi.org/legacy/
 
     - name: Publish distribution to PyPI
       uses: pypa/gh-action-pypi-publish@release/v1
       with:
-        password: ${{ secrets.PYPI_API_TOKEN }}
         repository_url: https://upload.pypi.org/legacy/
