### 2023-08-15 05b4f27
Aho-Corasick for multiple substring construction
