### 2023-08-15 d1b2c24
Add simple tests for Aho-Corasick
