+++ tests/test_dfa.py
@@ -2134,6 +2134,24 @@ class TestDFA(test_fa.TestFA):
             self.assertIn("nano", word)
 
     @params(True, False)
+    def test_contains_substrings(self, as_partial: bool) -> None:
+        input_symbols = {"a", "n", "o", "b"}
+        substring_dfa = DFA.from_substring(input_symbols, "nano")
+        substrings_dfa = DFA.from_substrings(input_symbols, ["nano"])
+        
+        self.assertEqual(substring_dfa, substrings_dfa)
+
+        substring_dfa = substring_dfa | DFA.from_substring(input_symbols, "banana")
+        substrings_dfa = DFA.from_substrings(input_symbols, ["banana", "nano"])
+        
+
+        self.assertEqual(substring_dfa, substrings_dfa)
+        
+        self.assertEqual(
+            ~substrings_dfa, DFA.from_substrings(input_symbols, ["banana", "nano"], contains=False)
+        )
+
+    @params(True, False)
     def test_contains_subsequence(self, as_partial: bool) -> None:
         """Should compute the minimal DFA accepting strings with the given
         subsequence"""
