### 2023-08-15 9a7d8ac
Type annotations
