+++ joss/paper.bib
@@ -73,25 +73,15 @@
   year={2009}
 }
 
-@article{Erickson23,
-  author       = {Jeff Erickson and
-                  Jason Xia and
-                  Eliot Wong Robson and
-                  Tue Do and
-                  Aidan Tzur Glickman and
-                  Zhuofan Jia and
-                  Eric Jin and
-                  Jiwon Lee and
-                  Patrick Lin and
-                  Steven Pan and
-                  Samuel Ruggerio and
-                  Tomoko Sakurayama and
-                  Andrew Yin and
-                  Yael Gertner and
-                  Brad Solomon},
-  title        = {Auto-graded scaffolding exercises for theoretical computer science},
-  year         = {2023},
-  publisher    = {ASEE Conferences}
+@INPROCEEDINGS{Erickson23
+author = "Jeff Erickson, Jason Xia, Eliot Wong Robson, Tue Do, Aidan Tzur Glickman, Zhuofan Jia, Eric Jin, Jiwon Lee, Patrick Lin, Steven Pan, Samuel Ruggerio, Tomoko Sakurayama, Andrew Yin, Yael Gertner, and Brad Solomon"
+title = "Auto-graded Scaffolding Exercises For Theoretical Computer Science"
+booktitle = "2023 ASEE Annual Conference \& Exposition"
+year = "2023"
+month = "June"
+address = "Baltimore , Maryland"
+publisher = "ASEE Conferences"
+note = {https://peer.asee.org/42347}
 }
 
 @misc{Johnson_2010,
