+++ joss/paper.bib
@@ -73,7 +73,7 @@
   year={2009}
 }
 
-@INPROCEEDINGS{Erickson23
+@INPROCEEDINGS{Erickson23,
 author = "Jeff Erickson, Jason Xia, Eliot Wong Robson, Tue Do, Aidan Tzur Glickman, Zhuofan Jia, Eric Jin, Jiwon Lee, Patrick Lin, Steven Pan, Samuel Ruggerio, Tomoko Sakurayama, Andrew Yin, Yael Gertner, and Brad Solomon"
 title = "Auto-graded Scaffolding Exercises For Theoretical Computer Science"
 booktitle = "2023 ASEE Annual Conference \& Exposition"
