### 2023-08-22 07f3840
More typo fixes in citation
