+++ joss/paper.bib
@@ -74,14 +74,14 @@
 }
 
 @INPROCEEDINGS{Erickson23,
-author = "Jeff Erickson, Jason Xia, Eliot Wong Robson, Tue Do, Aidan Tzur Glickman, Zhuofan Jia, Eric Jin, Jiwon Lee, Patrick Lin, Steven Pan, Samuel Ruggerio, Tomoko Sakurayama, Andrew Yin, Yael Gertner, and Brad Solomon"
-title = "Auto-graded Scaffolding Exercises For Theoretical Computer Science"
-booktitle = "2023 ASEE Annual Conference \& Exposition"
-year = "2023"
-month = "June"
-address = "Baltimore , Maryland"
-publisher = "ASEE Conferences"
-note = {https://peer.asee.org/42347}
+  author = {Jeff Erickson, Jason Xia, Eliot Wong Robson, Tue Do, Aidan Tzur Glickman, Zhuofan Jia, Eric Jin, Jiwon Lee, Patrick Lin, Steven Pan, Samuel Ruggerio, Tomoko Sakurayama, Andrew Yin, Yael Gertner, and Brad Solomon},
+  title = {Auto-graded Scaffolding Exercises For Theoretical Computer Science},
+  booktitle = {2023 ASEE Annual Conference \& Exposition},
+  year = {2023},
+  month = {June},
+  address = {Baltimore , Maryland},
+  publisher = {ASEE Conferences},
+  note = {https://peer.asee.org/42347}
 }
 
 @misc{Johnson_2010,
