### 2023-08-23 caleb531/automata#168 @mahfuz05062
Issue with installation for pip install 'automata-lib[visual]'
When I try to install with `visual`, I get this error. Is there anything else needed to install this? I understand that this is failing for `pygraphviz`, however, I am getting similar error in different OS (Linux, Mac).

pip install 'automata-lib[visual]'
Collecting automata-lib[visual]
  Obtaining dependency information for automata-lib[visual] from https://files.pythonhosted.org/packages/69/59/1ff43ba9ed01db2596477367896afea499023e2157ae6305d7737314c0c4/automata_lib-8.0.0-py3-none-any.whl.metadata
  Downloading automata_lib-8.0.0-py3-none-any.whl.metadata (4.4 kB)
Collecting networkx>=2.6.2 (from automata-lib[visual])
  Using cached networkx-3.1-py3-none-any.whl (2.1 MB)
Collecting frozendict>=2.3.4 (from automata-lib[visual])
  Downloading frozendict-2.3.8-cp39-cp39-manylinux_2_17_x86_64.manylinux2014_x86_64.whl (114 kB)
     ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 114.8/114.8 kB 700.2 kB/s eta 0:00:00
Collecting typing-extensions>=4.5.0 (from automata-lib[visual])
  Obtaining dependency information for typing-extensions>=4.5.0 from https://files.pythonhosted.org/packages/ec/6b/63cc3df74987c36fe26157ee12e09e8f9db4de771e0f3404263117e75b95/typing_extensions-4.7.1-py3-none-any.whl.metadata
  Downloading typing_extensions-4.7.1-py3-none-any.whl.metadata (3.1 kB)
Collecting cached-method>=0.1.0 (from automata-lib[visual])
  Using cached cached_method-0.1.0-py3-none-any.whl (4.2 kB)
Collecting coloraide>=1.8.2 (from automata-lib[visual])
  Obtaining dependency information for coloraide>=1.8.2 from https://files.pythonhosted.org/packages/98/13/b4fe026086ff9af45e690dcfffa811000ecb31e511cfba7358118fc5ac87/coloraide-2.9-py3-none-any.whl.metadata
  Downloading coloraide-2.9-py3-none-any.whl.metadata (4.0 kB)
Collecting pygraphviz>=1.10 (from automata-lib[visual])
  Using cached pygraphviz-1.11.zip (120 kB)
  Preparing metadata (setup.py) ... done
Downloading coloraide-2.9-py3-none-any.whl (234 kB)
   ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━ 234.0/234.0 kB 1.6 MB/s eta 0:00:00
Downloading typing_extensions-4.7.1-py3-none-any.whl (33 kB)
Using cached automata_lib-8.0.0-py3-none-any.whl (59 kB)
Building wheels for collected packages: pygraphviz
  Building wheel for pygraphviz (setup.py) ... error
  error: subprocess-exited-with-error

  × python setup.py bdist_wheel did not run successfully.
  │ exit code: 1
  ╰─> [55 lines of output]
      running bdist_wheel
      running build
      running build_py
      creating build
      creating build/lib.linux-x86_64-cpython-39
      creating build/lib.linux-x86_64-cpython-39/pygraphviz
      copying pygraphviz/testing.py -> build/lib.linux-x86_64-cpython-39/pygraphviz
      copying pygraphviz/__init__.py -> build/lib.linux-x86_64-cpython-39/pygraphviz
      copying pygraphviz/agraph.py -> build/lib.linux-x86_64-cpython-39/pygraphviz
      copying pygraphviz/graphviz.py -> build/lib.linux-x86_64-cpython-39/pygraphviz
      copying pygraphviz/scraper.py -> build/lib.linux-x86_64-cpython-39/pygraphviz
      creating build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_html.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_node_attributes.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/__init__.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_drawing.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_unicode.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_clear.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_repr_mimebundle.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_graph.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_attribute_defaults.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_close.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_scraper.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_string.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_layout.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_readwrite.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_subgraph.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      copying pygraphviz/tests/test_edge_attributes.py -> build/lib.linux-x86_64-cpython-39/pygraphviz/tests
      running egg_info
      writing pygraphviz.egg-info/PKG-INFO
      writing dependency_links to pygraphviz.egg-info/dependency_links.txt
      writing top-level names to pygraphviz.egg-info/top_level.txt
      reading manifest file 'pygraphviz.egg-info/SOURCES.txt'
      reading manifest template 'MANIFEST.in'
      warning: no files found matching '*.png' under directory 'doc'
      warning: no files found matching '*.txt' under directory 'doc'
      warning: no files found matching '*.css' under directory 'doc'
      warning: no previously-included files matching '*~' found anywhere in distribution
      warning: no previously-included files matching '*.pyc' found anywhere in distribution
      warning: no previously-included files matching '.svn' found anywhere in distribution
      no previously-included directories found matching 'doc/build'
      adding license file 'LICENSE'
      writing manifest file 'pygraphviz.egg-info/SOURCES.txt'
      copying pygraphviz/graphviz.i -> build/lib.linux-x86_64-cpython-39/pygraphviz
      copying pygraphviz/graphviz_wrap.c -> build/lib.linux-x86_64-cpython-39/pygraphviz
      running build_ext
      building 'pygraphviz._graphviz' extension
      creating build/temp.linux-x86_64-cpython-39
      creating build/temp.linux-x86_64-cpython-39/pygraphviz
      gcc -pthread -B /media/mahfuz/0E294B045A519BEC/anaconda3_05_21/envs/automata_test/compiler_compat -Wno-unused-result -Wsign-compare -DNDEBUG -O2 -Wall -fPIC -O2 -isystem /media/mahfuz/0E294B045A519BEC/anaconda3_05_21/envs/automata_test/include -I/media/mahfuz/0E294B045A519BEC/anaconda3_05_21/envs/automata_test/include -fPIC -O2 -isystem /media/mahfuz/0E294B045A519BEC/anaconda3_05_21/envs/automata_test/include -fPIC -DSWIG_PYTHON_STRICT_BYTE_CHAR -I/media/mahfuz/0E294B045A519BEC/anaconda3_05_21/envs/automata_test/include/python3.9 -c pygraphviz/graphviz_wrap.c -o build/temp.linux-x86_64-cpython-39/pygraphviz/graphviz_wrap.o
      pygraphviz/graphviz_wrap.c:3020:10: fatal error: graphviz/cgraph.h: No such file or directory
       #include "graphviz/cgraph.h"
                ^~~~~~~~~~~~~~~~~~~
      compilation terminated.
      error: command '/usr/bin/gcc' failed with exit code 1
      [end of output]

  ERROR: Failed building wheel for pygraphviz
  Running setup.py clean for pygraphviz
Failed to build pygraphviz
ERROR: Could not build wheels for pygraphviz, which is required to install pyproject.toml-based projects

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5759 review.
