### 2023-08-23 caleb531/automata#168 @mahfuz05062
I tried a couple of environments, some of those had anaconda installed, and some miniconda. But, yes, all of them had conda and the environments are created through conda create -n <env_name> python=<python_version> or similar.
