### 2023-08-23 caleb531/automata#168 @eliotwrobson
I believe it's possible to install graphviz directly in your anaconda environment, try `conda install -c anaconda graphviz` then reinstalling.
