### 2023-08-26 #5759 @Viech
I'm busier than anticipated these days, so I'll do the review in parts. I've only read the paper so far.

### Software paper

#### Summary

The last two sentences (research attention and related software) seem a better fit for the statement of need to me. Maybe conclude the summary instead by stating on a very high level the central feature(s) of automata?

l. 11: "are as follows" → "are the following" ?

#### Statement of need

Half of the statement is currently phrased in terms of the shortcomings of other packages, and these packages are not cited. I suggest to instead stress what makes automata stand out (i.e., mention that you *do* offer an extensive test suite or support for a wide range of automata) and then simply list and cite the alternative packages without necessarily going into detail of what they lack. It would also be good to state the scope and the target audience already in this section (as per JOSS guidelines).

#### The automata package

The suitability of Python as a platform for this application is something I'd rather put in the statement of need since in the current section I expect to learn something about automata in particular. In any case I would keep the discussion of Python as a language very brief and not include references to Rust or Cython unless these platforms relate to automata in a particular way.

I think you should give, as appropriate, a few more details or references to back the qualitative claims you make towards the end of the section. For instance, you could specify in what course or at which institution automata was used in education (maybe already cite Erickson et al. here to back the educational use?) or you could cite notable dependent projects or research papers to make the "wide adoption" more apparent. I would strike "with more [citations] to come as the package becomes more popular" as this is true for all scientific software.

l. 31: "of each automata" → "of each automaton" / "of the automata"
l. 31: I'm not sure what you mean with "built-in Python data structures"; to me this signifies things like lists and dicts?
l. 32: "easy-of-use" → "ease of use"
l. 50: "to the author’s knowledge" → "to the authors’ knowledge" ?
l. 50: "this is the only Python package implementing a number of the algorithms stated above": here I don't understand what "a number of" means; maybe replace with "either" or "all" or "more than x", as appropriate?

#### Example usage

I'd cut the two statements in lines 67 and 72 that you are using features of automata for the task; this is clear.

I feel that the pseudocode portrays the task more lengthy than it could be. You could consider allowing users to write just `from automata import DFA, NFA` and use that in the example. Maybe follow that up with `from string import ascii_lowercase as input_symbols` since the conversion to a `set` does not seem necessary. I would personally not break line after every argument to the factory methods as the calls are straightforward. You could call `NFA.edit_distance(input_symbols, reference_str="they", max_edit_distance=2)` with named keyword arguments and save the two lines where you define them. These are just ideas to present the library a bit more snappy; feel free to stick with what you have if you prefer it.

As a final line, state for clarity which words will be contained in `found_words`?

l. 66: "within the target edit distance" → "within a target edit distance" / "within some edit distance" ?
l. 71: "within the given edit distance of a reference string": this is repetitive w.r.t. l. 66.
l. 73: Not sure what "for later" means.
l. 74: "all of the words in the result" → "the words in the intersection of the accepted languages" ?

#### References

- The Aho et al. reference is the only reference in which the title is linked to a website (which is also not the publisher's).
- You abbreviate "J. Autom. Lang. Comb." but write "SIAM Journal on Computing" out.
- The Erickson et al., Møller, and Yingjie references are missing a source (no journal/publisher/URL/"personal communication"/etc. given).
- The Johnson reference seems badly formatted; maybe swap the blog's and the article's names? The blog name is also mentioned twice.
