### 2023-08-26 #5759 @eliotwrobson
@Viech thanks for the feedback so far! With regard to the citation formatting inconsistencies, is there a way to force all citations to follow consistent formatting? The bibliography was autogenerated, and I'm not sure how to make things like the abbreviations used consistent.
