### 2023-08-26 #5759 @Viech
> With regard to the citation formatting inconsistencies, is there a way to force all citations to follow consistent formatting? The bibliography was autogenerated, and I'm not sure how to make things like the abbreviations used consistent.

It looks like the journal abbreviation is hardcoded in your [.bib file](https://github.com/caleb531/automata/blob/develop/joss/paper.bib). You will probably need to fix this file by hand or with an editor like jabref (recommended) or kbibtex. Sourcing bibtex data from more than one place is bound to be inconsistent and the quality may vary quite a bit from source to source as well.
