### 2023-09-03 caleb531/automata#168 @caleb531
@mahfuz05062 Just checking in here—were you able to get `automata-lib[visual]` properly installed with @eliotwrobson's guidance? Or are you still having trouble?
