+++ joss/paper.md
@@ -62,7 +62,7 @@ Of note are some sophisticated and useful algorithms implemented in the package 
 
 - Thompson's algorithm for converting regular expressions to equivalent nondeterministic finite automata (NFA) [@AhoSU86].
 
-- Hopcroft's algorithm for DFA minimization [@Yingjie09].
+- Hopcroft's algorithm for DFA minimization [@Hopcroft71].
 
 - A specialized algorithm for directly constructing a state-minimal DFA accepting a given
 finite language [@mihov_schulz_2019].
