### 2023-09-03 f54e6a0
Addressed most reviewer comments
