+++ joss/paper.md
@@ -30,61 +30,53 @@ the set of all strings it accepts. Three important families of automata in incre
 2. Pushdown automata
 3. Turing machines
 
-These models are a core component of both computer science education and research, seeing applications in a wide variety of areas. In particular, the ability to manipulate finite-state automata within the context of a software package has seen attention from researchers in the past [@Sutner03]. Similar software has also included
-functionality for parsing regular expressions into their corresponding finite-state automata [@brics].
+The `automata` package facilitates working with these families by allowing simulation of reading input and higher-level manipulation
+of the corresponding languages using specialized algorithms.
 
 # Statement of need
 
-Although there are other packages in the Python software ecosystem that allow for working with
-various kinds of automata, they are often niche and lack things like a comprehensive test suite that
-allow for more rapid development. This leads to these packages being unable to adopt features that
-would be useful to researchers and students alike, such as sophisticated construction and manipulation
-algorithms. Moreover, Python is a popular tool for students and researchers, meaning the availability
-of a high-quality software package is likely to encourage the further exploration of these structures
-in the academic community.
+These models are a core component of both computer science education and research, seeing applications in a wide variety of areas.
+In particular, the ability to manipulate finite-state automata within the context of a software package has seen attention from
+researchers in the past [@Sutner03]. Similar software has also included
+functionality for parsing regular expressions into their corresponding finite-state automata [@brics].
+
+`automata` serves the demand for such a package in the Python software ecosystem. As a popular high-level language, Python enables
+significant flexibility and ease of use that of direct benefit to many users. The package includes a comprehensive test suite,
+support for modern language features (including type annotations), and has a large number of different automata,
+meeting the demands of users across a wide variety of use cases. In particular, the target audience
+is both researchers that wish to manipulate automata with more specialized algorithms and in educational contexts to
+reinforce understanding about how these models of computation function.
+
 
 # The `automata` package
 
-`automata` is a Python package for the manipulation and simulation of automata from the families listed above.
-The API is designed to mimic the formal mathematical description of each automaton using built-in Python data structures
-(such as sets and dicts).
-As a popular high-level language, Python enables greater flexibility and ease of use that is difficult
-to achieve with a low-level language (e.g., Rust). Algorithms in the package have been optimized for
-performance against benchmarks from tasks arising in research. In addition, Python allows for
-greater optimization by the integration of lower-level technologies (e.g., Cython), while still
-retaining the same high-level API, allowing for integration of more performant features as-needed by
-the user base. The package also has native display integration with Jupyter notebooks, enabling
-easy visualization.
+The API of the package is designed to mimic the formal mathematical description of each automaton using built-in Python data structures
+(such as sets and dicts). This is for ease of use by those that are unfamiliar with these structures, while also providing performance
+suitable for tasks arising in research. In particular, algorithms in the package have been optimized for
+performance on large automata, incorporating algorithmic optimizations such as only exploring the reachable set of states
+in the construction of a new deterministic finite-state automaton. The package also has native display integration with Jupyter
+notebooks, enabling easy visualization that allows students to interact with these structures in an exploratory manner.
 
 Of note are some sophisticated and useful algorithms implemented in the package for finite-state automata:
 
 - An optimized version of the Hopcroft-Karp algorithm to determine whether two deterministic finite automata (DFA) are equivalent [@AlmeidaMR10].
 
[…]
