+++ joss/paper.bib
@@ -115,3 +115,15 @@
     publisher = {Addison-Wesley Longman Publishing Co., Inc.},
     address = {USA}
 }
+
+@book{Sipser12,
+	series = {Introduction to the {Theory} of {Computation}},
+	title = {Introduction to the {Theory} of {Computation}},
+	isbn = {978-1-133-18781-3},
+	url = {https://books.google.com/books?id=4J1ZMAEACAAJ},
+	publisher = {Cengage Learning},
+	author = {Sipser, M.},
+	year = {2012},
+	lccn = {2012938665},
+    pages = {45-47}
+}
