### 2023-09-03 b8160c6
Added references
