+++ joss/paper.bib
@@ -127,3 +127,44 @@
 	lccn = {2012938665},
     pages = {45-47}
 }
+
+@article{Marschall11,
+	title = {Construction of minimal deterministic finite automata from biological motifs},
+	volume = {412},
+	issn = {0304-3975},
+	url = {https://www.sciencedirect.com/science/article/pii/S0304397510006948},
+	doi = {https://doi.org/10.1016/j.tcs.2010.12.003},
+	abstract = {Deterministic finite automata (DFAs) are constructed for various purposes in computational biology. Little attention, however, has been given to the efficient construction of minimal DFAs. In this article, we define simple non-deterministic finite automata (NFAs) and prove that the standard subset construction transforms NFAs of this type into minimal DFAs. Furthermore, we show how simple NFAs can be constructed from two types of pattern popular in bioinformatics, namely (sets of) generalized strings and (generalized) strings with a Hamming neighborhood.},
+	number = {8},
+	journal = {Theoretical Computer Science},
+	author = {Marschall, Tobias},
+	year = {2011},
+	keywords = {Consensus string, Deterministic finite automaton, Generalized string, Minimization, Motif},
+	pages = {922--930},
+}
+
+@article{Knuutila01,
+	title = {Re-describing an algorithm by {Hopcroft}},
+	volume = {250},
+	issn = {0304-3975},
+	url = {https://www.sciencedirect.com/science/article/pii/S0304397599001504},
+	doi = {https://doi.org/10.1016/S0304-3975(99)00150-4},
+	abstract = {J. Hopcroft introduced already in 1970 an O(nlogn)-time algorithm for minimizing a finite deterministic automaton of n states. Although the existence of the algorithm is widely known, its theoretical justification, correctness and running time analysis are not. We give here a tutorial reconstruction of Hopcroft's algorithm focusing on a firm theoretical basis, clear correctness proofs and a well-founded computational analysis. Our analysis reveals that if the size of the input alphabet m is not fixed, then Hopcroft's original algorithm does not run in time O(mnlogn) as is commonly believed in the literature. The O(mnlogn) holds, however, for the variation presented later by D. Gries and for a new variant given in this article. We also propose a new efficient routine for refining the equivalence classes constructed in the algorithm and suggest a computationally sound heuristics as an enhancement.},
+	number = {1},
+	journal = {Theoretical Computer Science},
+	author = {Knuutila, Timo},
+	year = {2001},
+	keywords = {Algorithms, Finite automata, Minimization},
+	pages = {333--363},
+}
+
+@ARTICLE{Xu16,
+  author={Xu, Chengcheng and Chen, Shuhui and Su, Jinshu and Yiu, S. M. and Hui, Lucas C. K.},
+  journal={IEEE Communications Surveys & Tutorials},
+  title={A Survey on Regular Expression Matching for Deep Packet Inspection: Applications, Algorithms, and Hardware Platforms},
+  year={2016},
+  volume={18},
+  number={4},
+  pages={2991-3029},
+  doi={10.1109/COMST.2016.2566669}
+}
