+++ joss/paper.md
@@ -102,8 +102,8 @@ A visualization of `target_words_dfa`, generated by the package in a Jupyter not
 is depicted in \autoref{fig:target_words_dfa}.
 
 Next, we construct an NFA recognizing all strings within a target edit distance of a fixed
-reference string. This construction can again be done with functions provided by the library,
-and we use the same alphabet as the DFA that was just created.
+reference string, and then immediately convert this to an equivalent DFA. The package provides
+builtin functions to make this construction easy, and we use the same alphabet as the DFA that was just created.
 
 ```python
 words_within_edit_distance_dfa = DFA.from_nfa(
