+++ joss/paper.bib
@@ -120,7 +120,6 @@
 	series = {Introduction to the {Theory} of {Computation}},
 	title = {Introduction to the {Theory} of {Computation}},
 	isbn = {978-1-133-18781-3},
-	url = {https://books.google.com/books?id=4J1ZMAEACAAJ},
 	publisher = {Cengage Learning},
 	author = {Sipser, M.},
 	year = {2012},
