### 2023-09-03 d3fb992
Address paper review comments (#169)
Made changes in response to JOSS review comments: https://github.com/openjournals/joss-reviews/issues/5759#issuecomment-1694400219
