+++ joss/paper.bib
@@ -17,7 +17,7 @@
   author =  {Anders M\o{}ller},
   title =   {dk.brics.automaton -- Finite-State Automata
              and Regular Expressions for {Java}},
-  note =    {\texttt{http://www.brics.dk/automaton/}},
+  url =    {http://www.brics.dk/automaton/},
   year =    2021
 }
 
@@ -26,7 +26,7 @@
                   Nelma Moreira and
                   Rog{\'{e}}rio Reis},
   title        = {Testing the Equivalence of Regular Languages},
-  journal      = {J. Autom. Lang. Comb.},
+  journal      = {Journal of Automata, Languages and Combinatorics},
   volume       = {15},
   number       = {1/2},
   pages        = {7--25},
@@ -51,27 +51,29 @@
   collection={Cambridge Tracts in Theoretical Computer Science}
 }
 
-@inbook{AhoSU86,
-  author       = {Alfred V. Aho and
-                  Ravi Sethi and
-                  Jeffrey D. Ullman},
-  title        = {Compilers: Principles, Techniques, and Tools},
-  series       = {Addison-Wesley series in computer science / World student series edition},
-  publisher    = {Addison-Wesley},
-  year         = {1986},
-  url          = {https://www.worldcat.org/oclc/12285707},
-  isbn         = {0-201-10088-6},
-  timestamp    = {Fri, 17 Jul 2020 16:12:45 +0200},
-  biburl       = {https://dblp.org/rec/books/aw/AhoSU86.bib},
-  bibsource    = {dblp computer science bibliography, https://dblp.org},
-  pages        = {159–163}
+@book{AhoSU86,
+    author = {Aho, Alfred V. and Lam, Monica S. and Sethi, Ravi and Ullman, Jeffrey D.},
+    title = {Compilers: Principles, Techniques, and Tools (2nd Edition)},
+    year = {2006},
+    isbn = {0321486811},
+    publisher = {Addison-Wesley Longman Publishing Co., Inc.},
+    address = {USA},
+    pages = {152-155}
 }
 
-@inproceedings{Yingjie09,
-  title={Describing an n log n algorithm for minimizing states in deterministic finite automaton},
-  author={Xu Yingjie},
-  year={2009}
+@incollection{Hopcroft71,
+    title = {AN n log n ALGORITHM FOR MINIMIZING STATES IN A FINITE AUTOMATON},
+    editor = {Zvi Kohavi and Azaria Paz},
+    booktitle = {Theory of Machines and Computations},
+    publisher = {Academic Press},
+    pages = {189-196},
+    year = {1971},
+    isbn = {978-0-12-417750-5},
+    doi = {https://doi.org/10.1016/B978-0-12-417750-5.50022-1},
+    url = {https://www.sciencedirect.com/science/article/pii/B9780124177505500221},
+    author = {John Hopcroft},
 }
+
 
 @INPROCEEDINGS{Erickson23,
   author = {Jeff Erickson, Jason Xia, Eliot Wong Robson, Tue Do, Aidan Tzur Glickman, Zhuofan Jia, Eric Jin, Jiwon Lee, Patrick Lin, Steven Pan, Samuel Ruggerio, Tomoko Sakurayama, Andrew Yin, Yael Gertner, and Brad Solomon},
@@ -81,13 +83,12 @@
   month = {June},
   address = {Baltimore , Maryland},
   publisher = {ASEE Conferences},
-  note = {https://peer.asee.org/42347}
+  url = {https://peer.asee.org/42347}
 }
 
 @misc{Johnson_2010,
   title={Nick’s Blog},
[…]
