### 2023-09-03 63afa8d
Fix invalid DOIs
