+++ joss/paper.bib
@@ -69,7 +69,7 @@
     pages = {189-196},
     year = {1971},
     isbn = {978-0-12-417750-5},
-    doi = {https://doi.org/10.1016/B978-0-12-417750-5.50022-1},
+    doi = {10.1016/B978-0-12-417750-5.50022-1},
     url = {https://www.sciencedirect.com/science/article/pii/B9780124177505500221},
     author = {John Hopcroft},
 }
@@ -132,7 +132,7 @@
 	volume = {412},
 	issn = {0304-3975},
 	url = {https://www.sciencedirect.com/science/article/pii/S0304397510006948},
-	doi = {https://doi.org/10.1016/j.tcs.2010.12.003},
+	doi = {10.1016/j.tcs.2010.12.003},
 	abstract = {Deterministic finite automata (DFAs) are constructed for various purposes in computational biology. Little attention, however, has been given to the efficient construction of minimal DFAs. In this article, we define simple non-deterministic finite automata (NFAs) and prove that the standard subset construction transforms NFAs of this type into minimal DFAs. Furthermore, we show how simple NFAs can be constructed from two types of pattern popular in bioinformatics, namely (sets of) generalized strings and (generalized) strings with a Hamming neighborhood.},
 	number = {8},
 	journal = {Theoretical Computer Science},
@@ -147,7 +147,7 @@
 	volume = {250},
 	issn = {0304-3975},
 	url = {https://www.sciencedirect.com/science/article/pii/S0304397599001504},
-	doi = {https://doi.org/10.1016/S0304-3975(99)00150-4},
+	doi = {10.1016/S0304-3975(99)00150-4},
 	abstract = {J. Hopcroft introduced already in 1970 an O(nlogn)-time algorithm for minimizing a finite deterministic automaton of n states. Although the existence of the algorithm is widely known, its theoretical justification, correctness and running time analysis are not. We give here a tutorial reconstruction of Hopcroft's algorithm focusing on a firm theoretical basis, clear correctness proofs and a well-founded computational analysis. Our analysis reveals that if the size of the input alphabet m is not fixed, then Hopcroft's original algorithm does not run in time O(mnlogn) as is commonly believed in the literature. The O(mnlogn) holds, however, for the variation presented later by D. Gries and for a new variant given in this article. We also propose a new efficient routine for refining the equivalence classes constructed in the algorithm and suggest a computationally sound heuristics as an enhancement.},
 	number = {1},
 	journal = {Theoretical Computer Science},
