### 2023-09-07 caleb531/automata#168 @mahfuz05062
Sorry, I was out on a trip last week, but yes, the suggestion worked. I am closing this issue.
