### 2023-09-07 #5759 @Viech
I went over the remaining points and found no issue with the software or repository. Both are in a really clean state! The repository and documentation landing pages (README.md and docs/README.md) are a bit bare bones but the actual documentation content is solid. I'm happy with the submission at this point but two things are missing as per the JOSS checklist:

- There is no statement-of-need kind of text in the *documentation*. Maybe extending one of the two README files in this direction would be an option?
- You don't describe the state of the field in the paper. Are there are any actively maintained competing or otherwise relevant packages that should be mentioned? Similar packages in other languages would also be a natural fit.

With respect to the paper updates, this looks good, except:

- I wouldn't start a section with "These models are", maybe write again "Automata are"?
- The Nick's blog reference now breaks out of the line. I don't know if this can be fixed; maybe @drvinceknight does?
