### 2023-09-08 #5759 @eliotwrobson
>     * There is no statement-of-need kind of text in the _documentation_. Maybe extending one of the two README files in this direction would be an option?

This should be straightforward, I'll add a short statement to the README.md lifted from the paper itself.

>     * You don't describe the state of the field in the paper. Are there are any actively maintained competing or otherwise relevant packages that should be mentioned? Similar packages in other languages would also be a natural fit.

I cited a couple of other similar packages, but I didn't include any detailed discussion of the features they provide. I'll update the paper with this section expanded and a couple of other references (along with addressing your latest round of comments).
