### 2023-09-08 e41cb97
Add pairwise helper function
