### 2023-09-08 4d99436
Paper bibliography update
