+++ joss/paper.bib
@@ -17,7 +17,7 @@
   author =  {Anders M\o{}ller},
   title =   {dk.brics.automaton -- Finite-State Automata
              and Regular Expressions for {Java}},
-  url =    {http://www.brics.dk/automaton/},
+  url =    {https://www.brics.dk/automaton/},
   year =    2021
 }
 
