### 2023-09-10 fa99bf6
Add discussion of other packages
