+++ joss/paper.bib
@@ -87,9 +87,9 @@
 }
 
 @misc{Johnson_2010,
-  title={Nick’s Blog},
+  title={Damn Cool Algorithms: Levenshtein Automata},
+  journal={Nick’s Blog},
   url={http://blog.notdot.net/2010/07/Damn-Cool-Algorithms-Levenshtein-Automata},
-  journal={Damn Cool Algorithms: Levenshtein Automata},
   author={Johnson, Nick},
   year={2010},
   month={Jul}
