### 2023-09-10 a4396b7
More paper edits (#171)
