+++ joss/paper.bib
@@ -87,7 +87,8 @@
 }
 
 @misc{Johnson_2010,
-  title={Nick’s Blog},
+  title={Damn Cool Algorithms: Levenshtein Automata},
+  journal={Nick’s Blog},
   url={http://blog.notdot.net/2010/07/Damn-Cool-Algorithms-Levenshtein-Automata},
   author={Johnson, Nick},
   year={2010},
@@ -166,4 +167,13 @@
   number={4},
   pages={2991-3029},
   doi={10.1109/COMST.2016.2566669}
+}
+
+@book{Rodger06,
+    author = {Rodger, Susan H.},
+    title = {JFLAP: An Interactive Formal Languages and Automata Package},
+    year = {2006},
+    isbn = {0763738344},
+    publisher = {Jones and Bartlett Publishers, Inc.},
+    address = {USA}
 }
