### 2023-09-10 #5759 @eliotwrobson
@Viech I've addressed all of your paper comments in your latest comment (including the reference URL being weird, that seemed to just randomly fix itself). The discussion of other relevant packages appears in the statement of need.

With respect to the statement of need in the documentation, I've created a PR with an updated `README.md` [here](https://github.com/caleb531/automata/pull/172). Feel free to leave comments. Once you think things are ready, I can go ahead and merge the changes.
