### 2023-09-10 b170f42
Update package description in README (#172)
