### 2023-09-10 caleb531/automata#173 @mahfuz05062
Running the test suites in the package
One criterion for the JOSS review is: "Are there automated tests or manual steps described so that the functionality of the software can be verified?"

Can we add a way to run the test suite automatically or at least provide a guideline to run the tests manually?

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5759 review.
