### 2023-09-10 caleb531/automata#174 @mahfuz05062
Guidelines for contribution
One criterion for the JOSS review is: "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

Can we include guidelines for potential contributors? Maybe a CONTRIBUTING.md file and direct the interested people there from the README.md file.

This issue is part of the https://github.com/openjournals/joss-reviews/issues/5759 review.
