### 2023-09-10 caleb531/automata#173 @eliotwrobson
So the test suite is run as part of the GitHub actions on the master branch, and can be run manually by using the `nose2` command. More details can be found here: https://github.com/caleb531/automata/blob/main/.github/CONTRIBUTING.md
