### 2023-09-10 caleb531/automata#174 @eliotwrobson
There is a CONTRIBUTING.md file in the `.github` folder: https://github.com/caleb531/automata/blob/main/.github/CONTRIBUTING.md

However, I do agree that it can be a bit hard to find. @mahfuz05062 I'll go ahead and add a link to the CONTRIBUTING.md from README.md on the front page.
