### 2023-09-10 3c84693
Added link to contributing guide
