+++ README.md
@@ -1,6 +1,6 @@
 # Automata
 
-*Copyright 2016-2023 Caleb Evans*  
+*Copyright 2016-2023 Caleb Evans*
 *Released under the MIT license*
 
 [![tests](https://github.com/caleb531/automata/actions/workflows/tests.yml/badge.svg)](https://github.com/caleb531/automata/actions/workflows/tests.yml)
@@ -61,3 +61,7 @@ of the GitHub repository.
 
 <!-- the below link must be an absolute URL to be functional in the PyPI README -->
 [docs]: https://github.com/caleb531/automata/blob/main/docs/README.md
+
+## Contributing
+
+Contributions are always welcome! Take a look at the [contributing guide](.github/CONTRIBUTING.md).
