### 2023-09-10 1c4a93a
Added link to contributing guide to README.md (#175)
* Added link to contributing guide

* Update README.md
