### 2023-10-08 #5759 @eliotwrobson
> could you make a tagged release and archive, and report the version number and archive DOI here. Please make sure the title and author list of the archive matches the paper.

Since I am not the repo owner, these steps will need to be done by @caleb531 (at least the archiving step on Zenodo). Will communicate with him to get this sorted out!
