### 2023-10-08 a568bab
Use same version number example in Release Checklist
Everywhere else in the file, v8 is used as the example, not v7.
