### 2023-10-08 #5759 @eliotwrobson
@drvinceknight the requested information is below:

- Version number: v8.1.0
- Archive DOI (Zenodo): 10.5281/zenodo.8419572
- Zenodo page link: https://zenodo.org/record/8419572

Please let me know if you need anything else!
