### 2023-10-09 #5759 @danielskatz
👋 @eliotwrobson - I'm the track editor for this submission, and I have a few very small changes to suggest - see https://github.com/caleb531/automata/pull/176. Please merge this, or let me know you disagree with, then we can continue to acceptance and publication
