### 2023-10-09 6dccd3f
Minor changes in paper and bib (#176)
* Update paper.md

* Update paper.bib
