+++ joss/paper.bib
@@ -62,7 +62,7 @@
 }
 
 @incollection{Hopcroft71,
-    title = {AN n log n ALGORITHM FOR MINIMIZING STATES IN A FINITE AUTOMATON},
+    title = {An n log n Algorithm for Minimizing States in a Finite Automaton},
     editor = {Zvi Kohavi and Azaria Paz},
     booktitle = {Theory of Machines and Computations},
     publisher = {Academic Press},
@@ -119,7 +119,7 @@
 
 @book{Sipser12,
 	series = {Introduction to the {Theory} of {Computation}},
-	title = {Introduction to the {Theory} of {Computation}},
+	title = {Introduction to the Theory of Computation},
 	isbn = {978-1-133-18781-3},
 	publisher = {Cengage Learning},
 	author = {Sipser, M.},
