### 2023-10-09 #5759 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
